move get_clause_str to user_story module
